go: Break up webserver into files
